Ensure sockets are opened with `O_CLOEXEC` (#603)

+++ library/runtime-ctrl/src/macosMain/kotlin/io/matthewnelson/kmp/tor/runtime/ctrl/internal/MacosPlatform.kt
@@ -25,7 +25,7 @@ import platform.posix.strcpy
 
 @OptIn(ExperimentalForeignApi::class)
 internal actual inline fun File.socketAddress(
-    family: UShort,
+    family: Int,
     block: (CValuesRef<sockaddr>, len: socklen_t) -> Unit,
 ) {
     cValue<sockaddr_un> {
